修正

+++ SuikomiKibun/src/player/playerteki.h
@@ -33,7 +33,7 @@ public:
 	void Draw();
 	void RenderScene();
 	void PlayerTekiMove(btVector3 pos);
-	void PlayerTekiResize(int size);
+	void PlayerTekiResize(double size);
 	void PlayerColorChange(int color_change);
 };
